CATTY-601 join all BlackboxParserTests in one file
