+++ src/CattyTests/XML/Parser/CatrobatLanguage0.95/XMLParserObjectTests095.swift
