+++ src/CattyTests/XML/Parser/CatrobatLanguage0.95/XMLParserUserVariableTests095.swift
