+++ src/CattyTests/XML/Serializer/XMLSerializerBlackBoxTests.swift
